[issue-929] split session state modules
